refactor: 패키지명 token으로 변경
