+++ src/main/java/com/snackgame/server/auth/token/domain/RefreshTokenRepository.java
@@ -1,10 +1,8 @@
-package com.snackgame.server.auth.jwt.domain;
+package com.snackgame.server.auth.token.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
-
-import com.snackgame.server.auth.jwt.domain.RefreshToken;
 
 @Repository
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
