+++ src/main/java/com/snackgame/server/auth/token/util/BearerTokenExtractor.java
@@ -1,4 +1,4 @@
-package com.snackgame.server.auth.jwt.util;
+package com.snackgame.server.auth.token.util;
 
 import java.util.Objects;
