+++ src/main/java/com/snackgame/server/member/domain/SnackgameMemberResolver.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import org.springframework.stereotype.Component;
 
-import com.snackgame.server.auth.jwt.support.MemberResolver;
+import com.snackgame.server.auth.token.support.MemberResolver;
 
 @Component
 public class SnackgameMemberResolver implements MemberResolver<Member> {
